Created metrics helper.
