+++ arangod/CMakeLists.txt
@@ -644,7 +644,7 @@ set(LIB_ARANGO_REPLICATION2_SOURCES
   RestHandler/RestLogHandler.cpp
   RestHandler/RestPrototypeStateHandler.cpp
   RestHandler/RestLogInternalHandler.cpp
-)
+        Replication2/MetricsHelper.cpp Replication2/MetricsHelper.h)
 
 set (LIB_ARANGO_METRICS_SOURCES
   Metrics/Builder.cpp
