+++ arangod/Replication2/ReplicatedLog/LogFollower.cpp
@@ -24,14 +24,15 @@
 #include "LogFollower.h"
 
 #include "Logger/LogContextKeys.h"
-#include "Metrics/Gauge.h"
 #include "Metrics/Counter.h"
+#include "Metrics/Gauge.h"
+#include "Replication2/Exceptions/ParticipantResignedException.h"
+#include "Replication2/MetricsHelper.h"
 #include "Replication2/ReplicatedLog/Algorithms.h"
 #include "Replication2/ReplicatedLog/LogStatus.h"
 #include "Replication2/ReplicatedLog/NetworkMessages.h"
 #include "Replication2/ReplicatedLog/ReplicatedLogIterator.h"
 #include "Replication2/ReplicatedLog/ReplicatedLogMetrics.h"
-#include "Replication2/Exceptions/ParticipantResignedException.h"
 
 #include <Basics/Exceptions.h>
 #include <Basics/Result.h>
